refactor(perl-dap): extract server lifecycle from lib facade (#0000)
