+++ crates/perl-dap/src/server/mod.rs
@@ -1,0 +1,9 @@
+//! DAP server configuration and lifecycle.
+
+mod config;
+mod lifecycle;
+mod mode;
+
+pub use config::DapConfig;
+pub use lifecycle::DapServer;
+pub use mode::DapMode;
